backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
numpy==1.26.0
pillow==10.0.1
pytesseract==0.3.10
opencv-python==4.8.1.78
pyaudio==0.2.13
openai-whisper==20231117
transformers==4.34.0
torch==2.1.0
pynput==1.7.6
websockets==11.0.3
python-dotenv==1.0.0
requests==2.31.0
